requirements.txt
amqp==5.1.1
asgiref==3.7.2
billiard==4.1.0
blinker==1.6.2
Brotli==1.1.0
celery==5.3.4
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
ConfigArgParse==1.7
Django==4.2.5
django-cors-headers==4.2.0
djangorestframework==3.14.0
Flask==3.0.0
Flask-BasicAuth==0.2.0
Flask-Cors==4.0.0
gevent==23.9.1
geventhttpclient==2.0.10
greenlet==3.0.0rc3
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
kombu==5.3.2
locust==2.16.1
MarkupSafe==2.1.3
msgpack==1.0.7
Pillow==10.0.1
prompt-toolkit==3.0.39
psutil==5.9.5
pycparser==2.21
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pywin32==306
pyzmq==25.1.1
redis==4.6.0
requests==2.31.0
roundrobin==0.0.4
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.5
vine==5.0.0
wcwidth==0.2.8
Werkzeug==3.0.0
zope.event==5.0
zope.interface==6.0
